# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FORJA API",
    "version": "0.1.0",
    "description": "Generado desde los schemas zod del backend (backend/src/openapi). No editar a mano — correr `npm run openapi:generate` después de cambiar una ruta."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de sesión emitido por /auth/login o /auth/register."
      },
      "clientIdLegacy": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "Modo legacy sin cuenta: identifica la instalación. Se ignora si viene Authorization Bearer válido."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true
          }
        }
      },
...
      "ExerciseListEntry": {
        "type": "object",
        "properties": {
          "exercise_name": {
            "type": "string"
          },
          "entries": {
            "type": "number"
          },
          "last_date": {
            "type": "string"
          }
        },
        "required": [
          "exercise_name",
          "entries",
          "last_date"
        ]
      },
      "PrsByWeekdayEntry": {
        "type": "object",
        "properties": {
          "weekday_index": {
            "type": "number"
          },
          "label": {
            "type": "string"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "weekday_index",
          "label",
          "count"
        ]
      },
      "LatestSet": {
        "type": "object",
        "properties": {
          "weight_kg": {
            "type": "number"
          },
          "reps": {
            "type": "number"
          },
          "set_number": {
            "type": "number"
          }
        },
        "required": [
          "weight_kg",
          "reps",
          "set_number"
        ]
      },
      "PersonalRecord": {
        "type": "object",
        "nullable": true,
        "properties": {
          "exercise_name": {
            "type": "string"
          },
          "best_weight_kg": {
            "type": "number"
          },
          "best_weight_date": {
            "type": "string"
          },
          "best_volume": {
            "type": "number"
          },
          "best_volume_date": {
            "type": "string"
          }
        },
        "required": [
          "exercise_name",
          "best_weight_kg",
          "best_weight_date",
          "best_volume",
          "best_volume_date"
        ]
      },
      "ProgressionSuggestion": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "fuerza",
              "hipertrofia",
              "mantenimiento"
            ]
          },
          "action": {
            "type": "string",
            "enum": [
              "subir_peso",
              "mantener",
              "bajar",
              "sin_datos"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "mode",
          "action",
          "reason"
        ]
      },
      "RestSuggestion": {
        "type": "object",
        "properties": {
          "rest_seconds": {
            "type": "number"
          },
          "zone": {
            "type": "string",
            "nullable": true,
            "enum": [
              "fuerza_maxima",
              "hipertrofia",
              "adaptacion",
              "potencia",
              "otra"
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "rest_seconds",
          "zone",
          "note"
        ]
      },
      "HistoryEntry": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "weight_kg": {
            "type": "number"
          },
          "reps": {
            "type": "number"
          },
          "set_number": {
            "type": "number"
          },
          "is_warmup": {
            "type": "boolean"
          },
          "tonelaje": {
            "type": "number",
            "nullable": true
          },
          "intensidad_pct": {
            "type": "number",
            "nullable": true
          },
          "zona": {
            "type": "string",
            "nullable": true,
            "enum": [
              "fuerza_maxima",
              "hipertrofia",
              "adaptacion",
              "potencia",
              "otra"
            ]
          }
        },
        "required": [
          "date",
          "weight_kg",
          "reps",
          "set_number",
          "is_warmup",
          "tonelaje",
          "intensidad_pct",
          "zona"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/history": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ejercicios con al menos un registro",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseListEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/prs-by-weekday": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "PRs históricos agrupados por día de la semana (siempre 7)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrsByWeekdayEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{exerciseName}/latest": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Series de trabajo del último día registrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LatestSet"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{exerciseName}/records": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Récord personal (o null)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalRecord"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{exerciseName}/suggestion": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "fuerza",
                "hipertrofia",
                "mantenimiento"
              ]
            },
            "required": false,
            "name": "mode",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sugerencia de progresión para el próximo set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionSuggestion"
                }
              }
            }
          },
          "400": {
            "description": "Sin modo elegido (ni en el perfil ni por query)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{exerciseName}/rest-suggestion": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "weight_kg",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Descanso sugerido según intensidad relativa al PR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestSuggestion"
                }
              }
            }
          },
          "400": {
            "description": "Falta ?weight_kg=",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{exerciseName}": {
      "get": {
        "tags": [
          "history"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Evolución completa del ejercicio",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}